requirements.txt
openapi-python-client==0.22.0
python-dotenv==1.0.1
